[Chore] #207 - Setting View, ViewController 분리
